Remove the unnecessary memory access counter, and simplify sprite evaluation

+++ NesCoreCpp/PpuSpritesState.h
@@ -5,7 +5,7 @@ struct PpuSpritesState
     bool largeSprites_{};
     uint16_t spritePatternBase_{};
 
-    uint8_t oamAddress_{};
+    uint32_t oamAddress_{};
     std::array<uint8_t, 256> oam_{};
 
     bool sprite0Hit_{};
